fix: err hadnling, code structure, saved articles
